Working on records and JDBC API

+++ src/application/model/User.java
@@ -5,22 +5,22 @@ public class User {
     /* ======================
         VARIABLES
        ======================*/
-    private int userID;
+    private long userID;
     private String userName;
     private String password;
 
 
-    public User(int userID, String userName, String password) {
+    public User(long userID, String userName, String password) {
         this.userID = userID;
         this.userName = userName;
         this.password = password;
     }
 
-    public int getUserID() {
+    public long getUserID() {
         return userID;
     }
 
-    public void setUserID(int userID) {
+    public void setUserID(long userID) {
         this.userID = userID;
     }
